리액티브 마이크로서비스 적용 - 스프링 리액터 적용 - 스프링 데이터 MongoDB를 사용한 논블로킹 영속성 적용 - 핵심 서비스 논블로킹 REST API화 (Product, Recommendation, Review) - 복합 서비스 논블로킹 REST API화 (Product-Composite) - 이벤트 기반 비동기 서비스 적용 - 스프링 클라우드 스트림 구성 적용 -

+++ microservices/review-service/src/main/java/com/jaguarcode/microservices/core/review/services/ReviewServiceImpl.java
@@ -2,17 +2,23 @@ package com.jaguarcode.microservices.core.review.services;
 
 import com.jaguarcode.microservices.core.review.persistence.ReviewEntity;
 import com.jaguarcode.microservices.core.review.persistence.ReviewRepository;
+import org.reactivestreams.Publisher;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.web.bind.annotation.RestController;
+import reactor.core.publisher.Flux;
+import reactor.core.scheduler.Scheduler;
 import com.jaguarcode.api.core.review.Review;
 import com.jaguarcode.api.core.review.ReviewService;
 import com.jaguarcode.util.exceptions.InvalidInputException;
 import com.jaguarcode.util.http.ServiceUtil;
 
 import java.util.List;
+import java.util.function.Supplier;
+
+import static java.util.logging.Level.FINE;
 
 @RestController
 public class ReviewServiceImpl implements ReviewService {
@@ -25,8 +31,11 @@ public class ReviewServiceImpl implements ReviewService {
 
     private final ServiceUtil serviceUtil;
 
+    private final Scheduler scheduler;
+
     @Autowired
-    public ReviewServiceImpl(ReviewRepository repository, ReviewMapper mapper, ServiceUtil serviceUtil) {
+    public ReviewServiceImpl(Scheduler scheduler, ReviewRepository repository, ReviewMapper mapper, ServiceUtil serviceUtil) {
+        this.scheduler = scheduler;
         this.repository = repository;
         this.mapper = mapper;
         this.serviceUtil = serviceUtil;
@@ -34,6 +43,9 @@ public class ReviewServiceImpl implements ReviewService {
 
     @Override
     public Review createReview(Review body) {
+
+        if (body.getProductId() < 1) throw new InvalidInputException("Invalid productId: " + body.getProductId());
+
         try {
             ReviewEntity entity = mapper.apiToEntity(body);
             ReviewEntity newEntity = repository.save(entity);
@@ -47,9 +59,16 @@ public class ReviewServiceImpl implements ReviewService {
     }
 
     @Override
-    public List<Review> getReviews(int productId) {
+    public Flux<Review> getReviews(int productId) {
 
         if (productId < 1) throw new InvalidInputException("Invalid productId: " + productId);
+
+        LOG.info("Will get reviews for product with id={}", productId);
+
+        return asyncFlux(() -> Flux.fromIterable(getByProductId(productId))).log(null, FINE);
+    }
+
+    protected List<Review> getByProductId(int productId) {
 
         List<ReviewEntity> entityList = repository.findByProductId(productId);
         List<Review> list = mapper.entityListToApiList(entityList);
@@ -62,7 +81,13 @@ public class ReviewServiceImpl implements ReviewService {
 
     @Override
     public void deleteReviews(int productId) {
+        if (productId < 1) throw new InvalidInputException("Invalid productId: " + productId);
+
         LOG.debug("deleteReviews: tries to delete reviews for the product with productId: {}", productId);
         repository.deleteAll(repository.findByProductId(productId));
+    }
+
+    private <T> Flux<T> asyncFlux(Supplier<Publisher<T>> publisherSupplier) {
+        return Flux.defer(publisherSupplier).subscribeOn(scheduler);
     }
 }
